requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.32
passlib==1.7.4
protobuf==3.20.3
pydantic==1.10.5
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
